schedule ui

+++ backend/controllers/schedule.controller.js
@@ -1,29 +1,31 @@
 const schedule = require('../models/schedule.model')
 
 const axios = require('axios')
-// exports.postSchedule = (req, res)=>{
-//     // console.log("abc")
-// }
-
-// exports.postSchedule = async (req, res) => {
-//     // model.create(req.body, (err, data) => {
-//     //   if (err) {
-//     //     res.status(400).send(err);
-//     //   } else {
-//     //     res.status(200).send(data);
-//     //   }
-//     // });
-//     res.send({a:"AA"})
-// };
-// const Record = require('../models/record.model')
 
 exports.postSchedule = (req, res) => {
-    console.log(req.body)
-    data = req.body
-    console.log(data)
+    const scheduleList = req.body
+    const dates = scheduleList.dates
+    // đổi chế độ tưới -> xóa hết chế độ cũ
+    schedule.find({type:scheduleList.type})
+    .then(data => {
+      if(data.length == 0) schedule.collection.deleteMany({})
+    })
+    dates.map((date) => {
+      const data = {
+        ...scheduleList,
+        dates:date.substring(0,10)
+      }
+      // insert nếu ngày đó chưa có trong db
+      schedule.find({dates:data.dates})
+      .then((sche) => {
+        if(sche.length == 0){
+          schedule.collection.insertOne(data)
+        }
+      })
+    })
     // schedule.collection.remove({})
-    schedule.collection.deleteMany({})
-    schedule.collection.insertOne(data)
+    // schedule.collection.deleteMany({})
+    // schedule.collection.insertOne(data)
 }
 
 
@@ -32,7 +34,7 @@ const postReq = () => {
   const headers = {
     'Content-Type': 'application/json',
     'Authorization': 'JWT fefege...',
-    'X-AIO-Key':'aio_gkEv861mJUppkg8eGL9yVwNOpe9C'
+    'X-AIO-Key':'aio_Ydss25hDmgz9CswD0YQxHhDA0uIA'
   }
   const data2 = {
     "datum":{"value":"ON"}
@@ -50,12 +52,11 @@ const postReq = () => {
     })
 }
 
-exports.getSchedule = (req,res) => {
-    schedule.find({
-      "type" : { "$in": ["custom", "monthly","weekly"] }
-    })
+exports.getSchedule = () => {
+    console.log("aa")
+    schedule.find({})
     .then(schedules => {
-      
+      console.log(schedules)
       // get current date, time
       let day = new Date()
       let hour = new Date().getHours();
@@ -64,50 +65,77 @@ exports.getSchedule = (req,res) => {
       minute = minute<10?"0"+minute.toString():minute.toString();
       let curTime = hour+":"+minute;
 
-      console.log(schedules[0].time === curTime)
+    //   console.log(schedules[0].time === curTime)
 
-      // lặp lại theo tuần
-      if(schedules[0].type == "weekly" && schedules[0].dates[0][day.toDateString().substring(0,3)]){
-        postReq()
+    //   // lặp lại theo tuần
+      if(schedules[0].type == "weekly"){
+        schedules.forEach((obj) => {
+          if(obj.dates === day.toDateString().substring(0,3)){
+            console.log(obj.dates)
+            console.log(day.toDateString().substring(0,3))
+            postReq()
+          }
+        })
+        
       } 
       // lặp lại theo tháng
       else if(schedules[0].type == "monthly"){
-        // lấy ra ngày nhỏ nhất
-        const arr = schedules[0].dates.sort()
-        let days = new Date(schedules[0].dates[0])
-        // so sáng với ngày hiện tại  
-        if(days.getDate() === day.getDate()){
-          postReq()
-        }
+        schedules.forEach((obj) => {
+          let days = new Date(obj.dates)
+          if(days.getDate() === day.getDate()){
+            postReq()
+          }
+        })
       }
       // lặp lại trong khoảng thời gian
       else if(schedules[0].type === "custom"){
-        // lấy ra ngày nhỏ nhất
-        const arr = schedules[0].dates.sort()
-        let days = new Date(schedules[0].dates[0])
-        days = days.toDateString();
-
-        if(day.toDateString() === days){
-          postReq()
-          // xóa ngày vừa tưới 
-          arr.shift()
-          schedule.collection.updateOne(
-            {},
-            {$set:{dates:arr}}
-          )
-        }
+        schedules.forEach((obj) => {
+          let days = new Date(obj.dates)
+          days = days.toDateString();
+          if(day.toDateString() === days){
+            console.log(days)
+            console.log(day.toDateString())
+            postReq()
+            schedule.collection.deleteOne({dates:schedules[0].dates})
+          }
+        })
       }
     })
     .catch(err => console.log(err))
-    res.status(200).send("Success")
+    // res.status(200).send("Success")
 }
 
 exports.getListSchedule = (req,res) => {
-  schedule.find({
-    "type" : { "$in": ["custom", "monthly","weekly"] }
-  })
+  schedule.find({})
   .then(schedules => {
     res.send(schedules)
 })
   .catch((err) => console.log(err))
+}
+
+exports.deleteById = (req,res) => {
+  console.log(req.params['id'])
+  schedule.find({_id:req.params['id']})
+  .then(sche => {
+    const query = {dates:sche[0].dates}
+    schedule.collection.deleteOne(query, function(err, obj) {
+      if (err) throw err;
+    });
+  })
+}
+
+exports.updateSchedule = (req,res) => {
+ 
+  schedule.find({_id:req.params['id']})
+  .then(sche => {
+    const data = req.body
+    const query = {dates:sche[0].dates}
+    const updateDoc = {
+      $set: {
+        time:data.time,
+        water:data.water
+      },
+    };
+    schedule.collection.updateOne(query,updateDoc)
+  })
 }
